[NRH-305] generalize donationpage navbar
