+++ spa/src/components/donationPage/pageContent/SHeaderBar.js
@@ -1,24 +1,13 @@
-import * as S from './SHeaderBar.styled';
-
 // Context
 import { usePage } from 'components/donationPage/DonationPage';
+
+// Children
+import DonationPageNavbar from 'components/donationPage/DonationPageNavbar';
 
 function SHeaderBar() {
   const { page } = usePage();
 
-  const getImageUrl = (img) => {
-    if (img instanceof File) {
-      return URL.createObjectURL(img);
-    } else return img;
-  };
-
-  return (
-    <S.SHeaderBar bgImg={getImageUrl(page?.header_bg_image)} data-testid="s-header-bar">
-      <a href={page?.header_link} target="_blank" rel="noreferrer noopener">
-        <S.SHeaderLogo src={getImageUrl(page?.header_logo)} />
-      </a>
-    </S.SHeaderBar>
-  );
+  return <DonationPageNavbar page={page} />;
 }
 
 SHeaderBar.type = 'SHeaderBar';
